% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdInStock.R
\name{prdInStock_read}
\alias{prdInStock_read}
\title{回料入库}
\usage{
prdInStock_read(
  file_name = "data-raw/回料粉碎记录表模板.xlsx",
  token = "9B6F803F-9D37-41A2-BDA0-70A7179AF0F3"
)
}
\arguments{
\item{file_name}{文件名}

\item{token}{口令}
}
\value{
返回值
}
\description{
回料入库
}
\examples{
prdInStock_read()
}
