% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdInStock.R
\name{prdInstockServer}
\alias{prdInstockServer}
\title{回料入库响应函数}
\usage{
prdInstockServer(
  input,
  output,
  session,
  token = "9B6F803F-9D37-41A2-BDA0-70A7179AF0F3"
)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{token}{口令}
}
\value{
返回值
}
\description{
回料入库响应函数
}
\examples{
prdInstockServer()
}
