% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdPickup.R
\name{prdPickup_query}
\alias{prdPickup_query}
\title{查询数据}
\usage{
prdPickup_query(
  token = "9B6F803F-9D37-41A2-BDA0-70A7179AF0F3",
  FStartDate = "2022-01-01",
  FEndDate = "2022-01-01",
  FCompanyName = "苏州赛普生物科技有限公司",
  FWorkshop = "生产部",
  FShiftName = "白班"
)
}
\arguments{
\item{token}{口令}

\item{FStartDate}{开始日期}

\item{FEndDate}{结束日期}

\item{FCompanyName}{公司名称}

\item{FWorkshop}{车间}

\item{FShiftName}{班次}
}
\value{

}
\description{
查询数据
}
