% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prdPickup.R
\name{prdPickup_read}
\alias{prdPickup_read}
\title{回料领用}
\usage{
prdPickup_read(
  file_name = "data-raw/回料使用记录表模板.xlsx",
  token = "9B6F803F-9D37-41A2-BDA0-70A7179AF0F3"
)
}
\arguments{
\item{file_name}{文件名}

\item{token}{口令}
}
\value{
返回值
}
\description{
回料领用
}
\examples{
prdPickup_read()
}
